requirements.txt
discord.py[voice]>=1.7.3
googletrans
discord_buttons_plugin
requests